Added Gridlex to Nav bar
Go to Gridles and use Control-F and put in "Media Query"
In this section, it will give me the size for each <a> to use.  Always look at the Size of the Page to see where the icons/text becomes overlapping and needs to be changed.
For example:  col-2_ xs-4 means it will go to 4 cols when the size of the screen goes to 568px